LayoutTests/http/tests/security/mixedContent/insecure-image-in-iframe-UpgradeMixedContent-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image.html requested insecure content from http://127.0.0.1:8080/security/resources/compass.jpg. This content was automatically upgraded and should be served over HTTPS.

This test loads a secure iframe that loads an insecure image. We should get a mixed content callback because the iframe is HTTPS but is displaying insecure content.
